fix: review refactor 2
