+++ app-mobile/src/AppRouter.tsx
@@ -9,7 +9,7 @@ import HighwayCodeDetailPage from './pages/highway-code/Detail';
 import HighwayCode from './pages/highway-code/HighwayCode';
 import TheoryTest from './pages/theory-test/TheoryTest';
 import HazardPerceptionDetail from './pages/theory-test/hazard-perception/Detail';
-import HazardPerceptionResult from './pages/theory-test/hazard-perception/Finished';
+import HazardPerceptionResult from './pages/theory-test/hazard-perception/Result';
 import HazardPerception from './pages/theory-test/hazard-perception/HazardPerception';
 import Learn from './pages/theory-test/learn/Learn';
 import PractiseTopic from './pages/theory-test/learn/practise-topics/PractiseTopic';
@@ -59,7 +59,7 @@ const Router: React.FC = () => {
                 />
 
                 <Route
-                    path="/theory-test/hazard-perception/:id/finished"
+                    path="/theory-test/hazard-perception/result"
                     render={() => <HazardPerceptionResult />}
                     exact={true}
                 />
